Focal length, sensor size calculations changes

+++ scripts/create_sensor_data.py
@@ -21,7 +21,7 @@ with urllib.request.urlopen(SENSOR_DATA) as response:
 
 #include <map>
 
-static std::map<std::string,float> sensorData = {
+static std::map<std::string,double> sensorData = {
 """)
         for k in d:
             focal = d[k]
